+++ vera/__init__.py
@@ -1,4 +1,4 @@
 """Vera: a programming language designed for LLMs."""
 
-__version__ = "0.0.120"
-version = "0.0.120"
+__version__ = "0.0.121"
+version = "0.0.121"